generic cf error page
